[chore] organise my project to suit a conventional project structure
