+++ v1/api/endpoints/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+endpoints_blueprint = Blueprint('endpoints', __name__)
+
+from .import views
